Latest version (csv connection + name display)
In this version, a connection between the frontend and the csv file has been established. This allows for the names of the plants to be directly fetched from the new csv file (with over 1000 plant species). What is left to do now is to provide accurate results or predict the occurrence of each plant species (when they are selected).